+++ RyuSocks/Auth/GSSAPI.cs
@@ -26,17 +26,19 @@ namespace RyuSocks.Auth
     [AuthMethodImpl(0x01)]
     public class GSSAPI : IProxyAuth
     {
+        public int WrapperLength => throw new NotImplementedException();
+
         public bool Authenticate(ReadOnlySpan<byte> incomingPacket, out ReadOnlySpan<byte> outgoingPacket)
         {
             throw new NotImplementedException();
         }
 
-        public ReadOnlySpan<byte> Wrap(ReadOnlySpan<byte> packet, ProxyEndpoint remoteEndpoint, out int wrapperLength)
+        public int Wrap(Span<byte> packet, int packetLength, ProxyEndpoint remoteEndpoint)
         {
             throw new NotImplementedException();
         }
 
-        public Span<byte> Unwrap(Span<byte> packet, out ProxyEndpoint remoteEndpoint, out int wrapperLength)
+        public int Unwrap(Span<byte> packet, int packetLength, out ProxyEndpoint remoteEndpoint)
         {
             throw new NotImplementedException();
         }