+++ RyuSocks/Commands/Client/UdpAssociateCommand.cs
@@ -54,24 +54,28 @@ namespace RyuSocks.Commands.Client
             Client.Send(request.AsSpan());
         }
 
-        public override ReadOnlySpan<byte> Wrap(ReadOnlySpan<byte> buffer, ProxyEndpoint remoteEndpoint, out int wrapperLength)
+        public override int Wrap(Span<byte> buffer, int packetLength, ProxyEndpoint remoteEndpoint)
         {
-            UdpPacket packet = new(remoteEndpoint, buffer.Length);
-            buffer.CopyTo(packet.UserData);
-            wrapperLength = packet.HeaderLength;
+            UdpPacket packet = new(remoteEndpoint, packetLength);
+            buffer[..packetLength].CopyTo(packet.UserData);
             packet.Validate();
 
-            return packet.AsSpan();
+            buffer.Clear();
+            packet.AsSpan().CopyTo(buffer);
+
+            return packet.Bytes.Length;
         }
 
-        public override Span<byte> Unwrap(Span<byte> buffer, out ProxyEndpoint remoteEndpoint, out int wrapperLength)
+        public override int Unwrap(Span<byte> buffer, int packetLength, out ProxyEndpoint remoteEndpoint)
         {
-            UdpPacket packet = new(buffer.ToArray());
+            UdpPacket packet = new(buffer[..packetLength].ToArray());
             remoteEndpoint = packet.ProxyEndpoint;
-            wrapperLength = packet.HeaderLength;
             packet.Validate();
 
-            return packet.UserData;
+            buffer.Clear();
+            packet.UserData.CopyTo(buffer);
+
+            return packet.Bytes.Length;
         }
 
         public override void ProcessResponse(CommandResponse response)